create feedback followere

+++ python/yellow_follower.py
@@ -39,7 +39,7 @@ while (True):
     else:
         cX, cY = 0, 0
 
-    e = (cX - cWidth/2) * 255 / cWidth*2
+    e = int((cX - cWidth/2) * 255 / cWidth*2)
 
     lCmd = np.clip(255 + e, 0, 255)
     rCmd = np.clip(255 - e, 0, 255)
